+++ Changelog.md
@@ -4,6 +4,8 @@
 - Fixed Refill Mouse Items deleting favorited stackable items
 - Fixed Pickup to Void Bag First moving favorited items inside the Void Bag
 - Made Move Items in Return to Previous Slot more granular
+- Fixed right-clicking items not marking items or removing marks
+- Fixed Return to Previous Slot on Consumption marking items when it should not
 
 ## v0.9.0.1
 - Added Move Items To Return to Previous Slot